Sort serviceUrls to put the first valid/non-broken one first in the list
